float into int

+++ tsp_cplex.cc
@@ -71,7 +71,7 @@ IloInt TWBuilder(const TSPTWDataDT &data, string filename) {
     const int size_missions_multipleTW = data.SizeMissionsMultipleTW();
     const int nbVehicle = data.NbVehicles();
     const vector<int> Capa = data.CapaVecs();
-    const vector<float> demand = data.Demands();
+    const vector<int> demand = data.Demands();
     const vector<int> duration = data.Durations();
     const vector<vector<int>> tw_start = data.TimeWindowStarts();
     const vector<vector<int>> tw_end = data.TimeWindowEnds();
